fix: makes doc dir before writing to it and adds to .gitignore (#426)
When writing docs, if you do not already have a `/src/docs` directory, the script would fail while attempting to write to it. Now it checks for that directory and creates one if needed. The reason this could happen is because that directory is ignored from git, and so unless you've previously written to it, you will not have this directory on your machine.