feat: local docs (#432)
Brings back the ability to preview docs from a local repository, bypassing the need to save docs as records in the database. Does by so querying the `payload` repository from your local file system and traversing the `docs` directory. This will output a json file containing all mdx markup which is converted to Lexical before rendering.