fix: ensures docs topic order is kept in sync (#434)
Docs topics are currently stored in two unique files which are easily confused and kept out of sync (see #433). One is a CommonJS file that is used for the `pnpm build:localDocs` script, and the other is a TypeScript file. The .cjs file is now generated from the .ts file when the script runs to ensure they always match, and now includes a comment block to ensure it is never directly edited.