chore: updates .env.example (#458)
The `.env.example` was lacking a few key variables which made it difficult to copy to `.env` and get up and running quickly. 

Namely, the `DATABASE_URI`, `PAYLOAD_SECRET`, and `BLOB_STORE_ID` which are required to boot the app. 

The `NEXT_PUBLIC_CMS_URL` variable is also no longer needed, as the CMS is now runs within same app as the frontend since migrating to v3.

I've also included stub paths for the local docs directory to make it easier to run `pnpm fetchDocs` from a fresh project.